azure-function/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
requests
pyairtable
Unidecode
python-dotenv
azure-storage-file-datalake 
azure-identity
outscraper
azure-functions-durable
jsonpickle